+++ frontend/.eslintignore
@@ -1,1 +1,3 @@
-static
+public/*
+dist/*
+node_modules/*
