+++ frontend/.prettierignore
@@ -1,0 +1,3 @@
+public/*
+dist/*
+node_modules/*
